Refactor string_edge_cases.rs into focused test modules
Split the 1,709-line string_edge_cases.rs file into 5 focused test modules
for better maintainability and navigation. All 116 tests preserved and passing.

## Changes

### New Structure
Created `crates/executor/tests/string_tests/` directory with modules:
- **case_conversion.rs** (174 lines, 13 tests) - UPPER, LOWER
- **length_functions.rs** (301 lines, 22 tests) - LENGTH, CHAR_LENGTH, OCTET_LENGTH
- **substring_operations.rs** (503 lines, 32 tests) - SUBSTRING, LEFT, RIGHT
- **search_functions.rs** (367 lines, 23 tests) - POSITION, INSTR, LOCATE
- **transformations.rs** (401 lines, 26 tests) - CONCAT, REPLACE, REVERSE, TRIM
- **mod.rs** (30 lines) - Module documentation and organization

### Test Organization
Each module includes:
- Comprehensive module documentation
- NULL handling tests
- Empty string edge cases
- Unicode/multi-byte UTF-8 character tests
- Boundary conditions
- Error conditions (wrong arg count/types)
- Type conversion tests (VARCHAR vs CHARACTER)

### Files Modified
- Removed: `string_edge_cases.rs` (1,709 lines)
- Added: `string_edge_case_tests.rs` (test harness)
- Added: `string_tests/` directory with 6 files (1,776 lines total)
- Updated: `Cargo.toml` to reference new test structure

### Verification
- All 116 tests pass ✅
- Clean module structure with proper imports
- Uses shared `create_test_evaluator()` from common module

Closes #618
