+++ crates/executor/Cargo.toml
@@ -36,3 +36,7 @@ path = "tests/test_numeric_edge_cases.rs"
 [[test]]
 name = "date_arithmetic_tests"
 path = "tests/date_arithmetic_tests.rs"
+
+[[test]]
+name = "string_edge_case_tests"
+path = "tests/string_edge_case_tests.rs"
