+++ crates/executor/tests/string_edge_case_tests.rs
@@ -1,0 +1,7 @@
+//! String function edge case test suite
+//!
+//! This test harness loads the comprehensive string function test modules.
+//! See the `string_tests` module for detailed organization and coverage.
+
+mod common;
+mod string_tests;
