--- crates/executor/tests/string_edge_cases.rs
@@ -1,1709 +1,0 @@
-//! Comprehensive edge case tests for string functions
-//!
-//! This test file covers:
-//! - NULL handling for all string functions
-//! - Unicode and multi-byte UTF-8 character handling
-//! - Boundary conditions and edge cases
-//! - Error conditions and invalid inputs
-//!
-//! Goal: Increase string.rs coverage from 45% to 80%+
-
-mod common;
-
-use common::create_test_evaluator;
-
-// ============================================================================
-// UPPER/LOWER Function Tests
-// ============================================================================
-
-#[test]
-fn test_upper_null() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "UPPER".to_string(),
-        args: vec![ast::Expression::Literal(types::SqlValue::Null)],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Null);
-}
-
-#[test]
-fn test_upper_empty() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "UPPER".to_string(),
-        args: vec![ast::Expression::Literal(types::SqlValue::Varchar("".to_string()))],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Varchar("".to_string()));
-}
-
-#[test]
-fn test_upper_unicode() {
-    let (evaluator, row) = create_test_evaluator();
-    // Test Greek, accented characters, emojis
-    let expr = ast::Expression::Function {
-        name: "UPPER".to_string(),
-        args: vec![ast::Expression::Literal(types::SqlValue::Varchar("café".to_string()))],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Varchar("CAFÉ".to_string()));
-}
-
-#[test]
-fn test_upper_mixed_case() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "UPPER".to_string(),
-        args: vec![ast::Expression::Literal(types::SqlValue::Varchar("HeLLo WoRLd".to_string()))],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Varchar("HELLO WORLD".to_string()));
-}
-
-#[test]
-fn test_upper_character_type() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "UPPER".to_string(),
-        args: vec![ast::Expression::Literal(types::SqlValue::Character("test".to_string()))],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Varchar("TEST".to_string()));
-}
-
-#[test]
-fn test_upper_wrong_arg_count() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "UPPER".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("test".to_string())),
-            ast::Expression::Literal(types::SqlValue::Varchar("extra".to_string())),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row);
-    assert!(result.is_err());
-}
-
-#[test]
-fn test_upper_wrong_type() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "UPPER".to_string(),
-        args: vec![ast::Expression::Literal(types::SqlValue::Integer(123))],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row);
-    assert!(result.is_err());
-}
-
-#[test]
-fn test_lower_null() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "LOWER".to_string(),
-        args: vec![ast::Expression::Literal(types::SqlValue::Null)],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Null);
-}
-
-#[test]
-fn test_lower_empty() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "LOWER".to_string(),
-        args: vec![ast::Expression::Literal(types::SqlValue::Varchar("".to_string()))],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Varchar("".to_string()));
-}
-
-#[test]
-fn test_lower_unicode() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "LOWER".to_string(),
-        args: vec![ast::Expression::Literal(types::SqlValue::Varchar("CAFÉ".to_string()))],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Varchar("café".to_string()));
-}
-
-// ============================================================================
-// SUBSTRING Edge Cases
-// ============================================================================
-
-#[test]
-fn test_substring_null_string() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "SUBSTRING".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Null),
-            ast::Expression::Literal(types::SqlValue::Integer(1)),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Null);
-}
-
-#[test]
-fn test_substring_null_start() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "SUBSTRING".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string())),
-            ast::Expression::Literal(types::SqlValue::Null),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Null);
-}
-
-#[test]
-fn test_substring_null_length() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "SUBSTRING".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string())),
-            ast::Expression::Literal(types::SqlValue::Integer(1)),
-            ast::Expression::Literal(types::SqlValue::Null),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Null);
-}
-
-#[test]
-fn test_substring_negative_start() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "SUBSTRING".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string())),
-            ast::Expression::Literal(types::SqlValue::Integer(-5)),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    // SQL:1999 treats negative/zero start as position 1
-    assert_eq!(result, types::SqlValue::Varchar("hello".to_string()));
-}
-
-#[test]
-fn test_substring_zero_start() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "SUBSTRING".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string())),
-            ast::Expression::Literal(types::SqlValue::Integer(0)),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Varchar("hello".to_string()));
-}
-
-#[test]
-fn test_substring_start_beyond_length() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "SUBSTRING".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string())),
-            ast::Expression::Literal(types::SqlValue::Integer(100)),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Varchar("".to_string()));
-}
-
-#[test]
-fn test_substring_zero_length() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "SUBSTRING".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string())),
-            ast::Expression::Literal(types::SqlValue::Integer(2)),
-            ast::Expression::Literal(types::SqlValue::Integer(0)),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Varchar("".to_string()));
-}
-
-#[test]
-fn test_substring_negative_length() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "SUBSTRING".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string())),
-            ast::Expression::Literal(types::SqlValue::Integer(2)),
-            ast::Expression::Literal(types::SqlValue::Integer(-3)),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Varchar("".to_string()));
-}
-
-#[test]
-fn test_substring_multibyte_characters() {
-    let (evaluator, row) = create_test_evaluator();
-    // "café" is 4 characters but 5 bytes in UTF-8
-    let expr = ast::Expression::Function {
-        name: "SUBSTRING".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("café".to_string())),
-            ast::Expression::Literal(types::SqlValue::Integer(1)),
-            ast::Expression::Literal(types::SqlValue::Integer(3)),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Varchar("caf".to_string()));
-}
-
-#[test]
-fn test_substring_empty_string() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "SUBSTRING".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("".to_string())),
-            ast::Expression::Literal(types::SqlValue::Integer(1)),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Varchar("".to_string()));
-}
-
-// ============================================================================
-// CHAR_LENGTH and OCTET_LENGTH Tests
-// ============================================================================
-
-#[test]
-fn test_char_length_null() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "CHAR_LENGTH".to_string(),
-        args: vec![ast::Expression::Literal(types::SqlValue::Null)],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Null);
-}
-
-#[test]
-fn test_char_length_empty() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "CHAR_LENGTH".to_string(),
-        args: vec![ast::Expression::Literal(types::SqlValue::Varchar("".to_string()))],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Integer(0));
-}
-
-#[test]
-fn test_char_length_multibyte() {
-    let (evaluator, row) = create_test_evaluator();
-    // "café" is 4 characters but 5 bytes
-    let expr = ast::Expression::Function {
-        name: "CHAR_LENGTH".to_string(),
-        args: vec![ast::Expression::Literal(types::SqlValue::Varchar("café".to_string()))],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Integer(4));
-}
-
-#[test]
-fn test_char_length_using_characters() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "CHAR_LENGTH".to_string(),
-        args: vec![ast::Expression::Literal(types::SqlValue::Varchar("café".to_string()))],
-        character_unit: Some(ast::CharacterUnit::Characters),
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Integer(4));
-}
-
-#[test]
-fn test_char_length_using_octets() {
-    let (evaluator, row) = create_test_evaluator();
-    // "café" is 5 bytes in UTF-8 (c=1, a=1, f=1, é=2)
-    let expr = ast::Expression::Function {
-        name: "CHAR_LENGTH".to_string(),
-        args: vec![ast::Expression::Literal(types::SqlValue::Varchar("café".to_string()))],
-        character_unit: Some(ast::CharacterUnit::Octets),
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Integer(5));
-}
-
-#[test]
-fn test_octet_length_null() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "OCTET_LENGTH".to_string(),
-        args: vec![ast::Expression::Literal(types::SqlValue::Null)],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Null);
-}
-
-#[test]
-fn test_octet_length_empty() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "OCTET_LENGTH".to_string(),
-        args: vec![ast::Expression::Literal(types::SqlValue::Varchar("".to_string()))],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Integer(0));
-}
-
-#[test]
-fn test_octet_length_multibyte() {
-    let (evaluator, row) = create_test_evaluator();
-    // "café" is 5 bytes in UTF-8
-    let expr = ast::Expression::Function {
-        name: "OCTET_LENGTH".to_string(),
-        args: vec![ast::Expression::Literal(types::SqlValue::Varchar("café".to_string()))],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Integer(5));
-}
-
-#[test]
-fn test_octet_length_vs_char_length() {
-    let (evaluator, row) = create_test_evaluator();
-
-    // ASCII: same count
-    let ascii_expr = ast::Expression::Function {
-        name: "CHAR_LENGTH".to_string(),
-        args: vec![ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string()))],
-        character_unit: None,
-    };
-    let char_result = evaluator.eval(&ascii_expr, &row).unwrap();
-
-    let octet_expr = ast::Expression::Function {
-        name: "OCTET_LENGTH".to_string(),
-        args: vec![ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string()))],
-        character_unit: None,
-    };
-    let octet_result = evaluator.eval(&octet_expr, &row).unwrap();
-
-    assert_eq!(char_result, types::SqlValue::Integer(5));
-    assert_eq!(octet_result, types::SqlValue::Integer(5));
-}
-
-// ============================================================================
-// CONCAT Tests
-// ============================================================================
-
-#[test]
-fn test_concat_null_propagation() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "CONCAT".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string())),
-            ast::Expression::Literal(types::SqlValue::Null),
-            ast::Expression::Literal(types::SqlValue::Varchar("world".to_string())),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Null);
-}
-
-#[test]
-fn test_concat_empty_strings() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "CONCAT".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("".to_string())),
-            ast::Expression::Literal(types::SqlValue::Varchar("".to_string())),
-            ast::Expression::Literal(types::SqlValue::Varchar("".to_string())),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Varchar("".to_string()));
-}
-
-#[test]
-fn test_concat_single_arg() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "CONCAT".to_string(),
-        args: vec![ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string()))],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Varchar("hello".to_string()));
-}
-
-#[test]
-fn test_concat_many_args() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "CONCAT".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("a".to_string())),
-            ast::Expression::Literal(types::SqlValue::Varchar("b".to_string())),
-            ast::Expression::Literal(types::SqlValue::Varchar("c".to_string())),
-            ast::Expression::Literal(types::SqlValue::Varchar("d".to_string())),
-            ast::Expression::Literal(types::SqlValue::Varchar("e".to_string())),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Varchar("abcde".to_string()));
-}
-
-#[test]
-fn test_concat_integer_conversion() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "CONCAT".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("ID:".to_string())),
-            ast::Expression::Literal(types::SqlValue::Integer(123)),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Varchar("ID:123".to_string()));
-}
-
-// ============================================================================
-// POSITION Tests
-// ============================================================================
-
-#[test]
-fn test_position_null() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "POSITION".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Null),
-            ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string())),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Null);
-}
-
-#[test]
-fn test_position_not_found() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "POSITION".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("xyz".to_string())),
-            ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string())),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Integer(0));
-}
-
-#[test]
-fn test_position_found() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "POSITION".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("lo".to_string())),
-            ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string())),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Integer(4)); // 1-indexed
-}
-
-#[test]
-fn test_position_empty_needle() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "POSITION".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("".to_string())),
-            ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string())),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    // Empty string is found at position 1
-    assert_eq!(result, types::SqlValue::Integer(1));
-}
-
-#[test]
-fn test_position_multiple_occurrences() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "POSITION".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("l".to_string())),
-            ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string())),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    // Returns first occurrence (position 3, 1-indexed)
-    assert_eq!(result, types::SqlValue::Integer(3));
-}
-
-// ============================================================================
-// REPLACE Tests
-// ============================================================================
-
-#[test]
-fn test_replace_null() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "REPLACE".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Null),
-            ast::Expression::Literal(types::SqlValue::Varchar("a".to_string())),
-            ast::Expression::Literal(types::SqlValue::Varchar("b".to_string())),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Null);
-}
-
-#[test]
-fn test_replace_multiple_occurrences() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "REPLACE".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("hello hello".to_string())),
-            ast::Expression::Literal(types::SqlValue::Varchar("ll".to_string())),
-            ast::Expression::Literal(types::SqlValue::Varchar("rr".to_string())),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Varchar("herro herro".to_string()));
-}
-
-#[test]
-fn test_replace_empty_search() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "REPLACE".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string())),
-            ast::Expression::Literal(types::SqlValue::Varchar("".to_string())),
-            ast::Expression::Literal(types::SqlValue::Varchar("x".to_string())),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    // Empty string search inserts replacement between every character
-    assert_eq!(result, types::SqlValue::Varchar("xhxexlxlxox".to_string()));
-}
-
-#[test]
-fn test_replace_empty_replacement() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "REPLACE".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string())),
-            ast::Expression::Literal(types::SqlValue::Varchar("l".to_string())),
-            ast::Expression::Literal(types::SqlValue::Varchar("".to_string())),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Varchar("heo".to_string()));
-}
-
-#[test]
-fn test_replace_not_found() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "REPLACE".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string())),
-            ast::Expression::Literal(types::SqlValue::Varchar("xyz".to_string())),
-            ast::Expression::Literal(types::SqlValue::Varchar("abc".to_string())),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Varchar("hello".to_string()));
-}
-
-// ============================================================================
-// LEFT/RIGHT Tests
-// ============================================================================
-
-#[test]
-fn test_left_null() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "LEFT".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Null),
-            ast::Expression::Literal(types::SqlValue::Integer(3)),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Null);
-}
-
-#[test]
-fn test_left_negative_length() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "LEFT".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string())),
-            ast::Expression::Literal(types::SqlValue::Integer(-5)),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Varchar("".to_string()));
-}
-
-#[test]
-fn test_left_zero_length() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "LEFT".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string())),
-            ast::Expression::Literal(types::SqlValue::Integer(0)),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Varchar("".to_string()));
-}
-
-#[test]
-fn test_left_length_exceeds_string() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "LEFT".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string())),
-            ast::Expression::Literal(types::SqlValue::Integer(100)),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Varchar("hello".to_string()));
-}
-
-#[test]
-fn test_left_multibyte() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "LEFT".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("café".to_string())),
-            ast::Expression::Literal(types::SqlValue::Integer(3)),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Varchar("caf".to_string()));
-}
-
-#[test]
-fn test_right_null() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "RIGHT".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Null),
-            ast::Expression::Literal(types::SqlValue::Integer(3)),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Null);
-}
-
-#[test]
-fn test_right_negative_length() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "RIGHT".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string())),
-            ast::Expression::Literal(types::SqlValue::Integer(-5)),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Varchar("".to_string()));
-}
-
-#[test]
-fn test_right_zero_length() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "RIGHT".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string())),
-            ast::Expression::Literal(types::SqlValue::Integer(0)),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Varchar("".to_string()));
-}
-
-#[test]
-fn test_right_length_exceeds_string() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "RIGHT".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string())),
-            ast::Expression::Literal(types::SqlValue::Integer(100)),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Varchar("hello".to_string()));
-}
-
-#[test]
-fn test_right_multibyte() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "RIGHT".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("café".to_string())),
-            ast::Expression::Literal(types::SqlValue::Integer(3)),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Varchar("afé".to_string()));
-}
-
-// ============================================================================
-// INSTR/LOCATE Tests
-// ============================================================================
-
-#[test]
-fn test_instr_null() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "INSTR".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Null),
-            ast::Expression::Literal(types::SqlValue::Varchar("lo".to_string())),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Null);
-}
-
-#[test]
-fn test_instr_not_found() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "INSTR".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string())),
-            ast::Expression::Literal(types::SqlValue::Varchar("xyz".to_string())),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Integer(0));
-}
-
-#[test]
-fn test_instr_found() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "INSTR".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string())),
-            ast::Expression::Literal(types::SqlValue::Varchar("ll".to_string())),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Integer(3)); // 1-indexed
-}
-
-#[test]
-fn test_locate_null() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "LOCATE".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Null),
-            ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string())),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Null);
-}
-
-#[test]
-fn test_locate_not_found() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "LOCATE".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("xyz".to_string())),
-            ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string())),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Integer(0));
-}
-
-#[test]
-fn test_locate_with_start_position() {
-    let (evaluator, row) = create_test_evaluator();
-    // Find second occurrence of 'l' in 'hello'
-    let expr = ast::Expression::Function {
-        name: "LOCATE".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("l".to_string())),
-            ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string())),
-            ast::Expression::Literal(types::SqlValue::Integer(4)), // Start after first 'l'
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Integer(4)); // Found at position 4
-}
-
-#[test]
-fn test_locate_start_beyond_length() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "LOCATE".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("l".to_string())),
-            ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string())),
-            ast::Expression::Literal(types::SqlValue::Integer(100)),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Integer(0)); // Not found
-}
-
-// ============================================================================
-// REVERSE Tests
-// ============================================================================
-
-#[test]
-fn test_reverse_null() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "REVERSE".to_string(),
-        args: vec![ast::Expression::Literal(types::SqlValue::Null)],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Null);
-}
-
-#[test]
-fn test_reverse_empty() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "REVERSE".to_string(),
-        args: vec![ast::Expression::Literal(types::SqlValue::Varchar("".to_string()))],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Varchar("".to_string()));
-}
-
-#[test]
-fn test_reverse_single_char() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "REVERSE".to_string(),
-        args: vec![ast::Expression::Literal(types::SqlValue::Varchar("a".to_string()))],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Varchar("a".to_string()));
-}
-
-#[test]
-fn test_reverse_basic() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "REVERSE".to_string(),
-        args: vec![ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string()))],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Varchar("olleh".to_string()));
-}
-
-#[test]
-fn test_reverse_multibyte() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "REVERSE".to_string(),
-        args: vec![ast::Expression::Literal(types::SqlValue::Varchar("café".to_string()))],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Varchar("éfac".to_string()));
-}
-
-// ============================================================================
-// LENGTH Tests
-// ============================================================================
-
-#[test]
-fn test_length_null() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "LENGTH".to_string(),
-        args: vec![ast::Expression::Literal(types::SqlValue::Null)],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Null);
-}
-
-#[test]
-fn test_length_empty() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "LENGTH".to_string(),
-        args: vec![ast::Expression::Literal(types::SqlValue::Varchar("".to_string()))],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Integer(0));
-}
-
-#[test]
-fn test_length_basic() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "LENGTH".to_string(),
-        args: vec![ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string()))],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Integer(5));
-}
-
-#[test]
-fn test_length_multibyte() {
-    let (evaluator, row) = create_test_evaluator();
-    // LENGTH returns byte count (unlike CHAR_LENGTH)
-    let expr = ast::Expression::Function {
-        name: "LENGTH".to_string(),
-        args: vec![ast::Expression::Literal(types::SqlValue::Varchar("café".to_string()))],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    // "café" is 5 bytes in UTF-8
-    assert_eq!(result, types::SqlValue::Integer(5));
-}
-
-// ============================================================================
-// TRIM Advanced Tests (NULL removal character)
-// ============================================================================
-
-#[test]
-fn test_trim_null_removal_char() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Trim {
-        position: Some(ast::TrimPosition::Both),
-        removal_char: Some(Box::new(ast::Expression::Literal(types::SqlValue::Null))),
-        string: Box::new(ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string()))),
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Null);
-}
-
-#[test]
-fn test_trim_null_string() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Trim {
-        position: None,
-        removal_char: None,
-        string: Box::new(ast::Expression::Literal(types::SqlValue::Null)),
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Null);
-}
-
-#[test]
-fn test_trim_custom_char_multibyte() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Trim {
-        position: Some(ast::TrimPosition::Both),
-        removal_char: Some(Box::new(ast::Expression::Literal(types::SqlValue::Varchar(
-            "x".to_string(),
-        )))),
-        string: Box::new(ast::Expression::Literal(types::SqlValue::Varchar(
-            "xxxhelloxxx".to_string(),
-        ))),
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Varchar("hello".to_string()));
-}
-
-// ============================================================================
-// Additional Error Condition Tests
-// ============================================================================
-
-#[test]
-fn test_substring_wrong_arg_count_too_few() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "SUBSTRING".to_string(),
-        args: vec![ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string()))],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row);
-    assert!(result.is_err());
-}
-
-#[test]
-fn test_substring_wrong_arg_count_too_many() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "SUBSTRING".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string())),
-            ast::Expression::Literal(types::SqlValue::Integer(1)),
-            ast::Expression::Literal(types::SqlValue::Integer(2)),
-            ast::Expression::Literal(types::SqlValue::Integer(3)),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row);
-    assert!(result.is_err());
-}
-
-#[test]
-fn test_substring_wrong_type_string() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "SUBSTRING".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Integer(123)),
-            ast::Expression::Literal(types::SqlValue::Integer(1)),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row);
-    assert!(result.is_err());
-}
-
-#[test]
-fn test_substring_wrong_type_start() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "SUBSTRING".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string())),
-            ast::Expression::Literal(types::SqlValue::Varchar("one".to_string())),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row);
-    assert!(result.is_err());
-}
-
-#[test]
-fn test_substring_wrong_type_length() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "SUBSTRING".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string())),
-            ast::Expression::Literal(types::SqlValue::Integer(1)),
-            ast::Expression::Literal(types::SqlValue::Varchar("two".to_string())),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row);
-    assert!(result.is_err());
-}
-
-#[test]
-fn test_concat_no_args() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "CONCAT".to_string(),
-        args: vec![],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row);
-    assert!(result.is_err());
-}
-
-#[test]
-fn test_position_wrong_arg_count() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "POSITION".to_string(),
-        args: vec![ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string()))],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row);
-    assert!(result.is_err());
-}
-
-#[test]
-fn test_position_wrong_type() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "POSITION".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Integer(123)),
-            ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string())),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row);
-    assert!(result.is_err());
-}
-
-#[test]
-fn test_replace_wrong_arg_count() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "REPLACE".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string())),
-            ast::Expression::Literal(types::SqlValue::Varchar("l".to_string())),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row);
-    assert!(result.is_err());
-}
-
-#[test]
-fn test_replace_wrong_type() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "REPLACE".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string())),
-            ast::Expression::Literal(types::SqlValue::Integer(123)),
-            ast::Expression::Literal(types::SqlValue::Varchar("x".to_string())),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row);
-    assert!(result.is_err());
-}
-
-#[test]
-fn test_left_wrong_arg_count() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "LEFT".to_string(),
-        args: vec![ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string()))],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row);
-    assert!(result.is_err());
-}
-
-#[test]
-fn test_left_wrong_type_string() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "LEFT".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Integer(123)),
-            ast::Expression::Literal(types::SqlValue::Integer(3)),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row);
-    assert!(result.is_err());
-}
-
-#[test]
-fn test_left_wrong_type_length() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "LEFT".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string())),
-            ast::Expression::Literal(types::SqlValue::Varchar("three".to_string())),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row);
-    assert!(result.is_err());
-}
-
-#[test]
-fn test_right_wrong_arg_count() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "RIGHT".to_string(),
-        args: vec![ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string()))],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row);
-    assert!(result.is_err());
-}
-
-#[test]
-fn test_right_wrong_type() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "RIGHT".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Integer(123)),
-            ast::Expression::Literal(types::SqlValue::Integer(3)),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row);
-    assert!(result.is_err());
-}
-
-#[test]
-fn test_instr_wrong_arg_count() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "INSTR".to_string(),
-        args: vec![ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string()))],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row);
-    assert!(result.is_err());
-}
-
-#[test]
-fn test_instr_wrong_type() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "INSTR".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Integer(123)),
-            ast::Expression::Literal(types::SqlValue::Varchar("l".to_string())),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row);
-    assert!(result.is_err());
-}
-
-#[test]
-fn test_locate_wrong_arg_count() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "LOCATE".to_string(),
-        args: vec![ast::Expression::Literal(types::SqlValue::Varchar("l".to_string()))],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row);
-    assert!(result.is_err());
-}
-
-#[test]
-fn test_locate_wrong_type_needle() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "LOCATE".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Integer(123)),
-            ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string())),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row);
-    assert!(result.is_err());
-}
-
-#[test]
-fn test_locate_wrong_type_start() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "LOCATE".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("l".to_string())),
-            ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string())),
-            ast::Expression::Literal(types::SqlValue::Varchar("one".to_string())),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row);
-    assert!(result.is_err());
-}
-
-#[test]
-fn test_locate_null_start() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "LOCATE".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("l".to_string())),
-            ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string())),
-            ast::Expression::Literal(types::SqlValue::Null),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Null);
-}
-
-#[test]
-fn test_reverse_wrong_arg_count() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "REVERSE".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string())),
-            ast::Expression::Literal(types::SqlValue::Varchar("extra".to_string())),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row);
-    assert!(result.is_err());
-}
-
-#[test]
-fn test_reverse_wrong_type() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "REVERSE".to_string(),
-        args: vec![ast::Expression::Literal(types::SqlValue::Integer(123))],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row);
-    assert!(result.is_err());
-}
-
-#[test]
-fn test_length_wrong_arg_count() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "LENGTH".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string())),
-            ast::Expression::Literal(types::SqlValue::Varchar("extra".to_string())),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row);
-    assert!(result.is_err());
-}
-
-#[test]
-fn test_length_wrong_type() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "LENGTH".to_string(),
-        args: vec![ast::Expression::Literal(types::SqlValue::Integer(123))],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row);
-    assert!(result.is_err());
-}
-
-#[test]
-fn test_char_length_wrong_arg_count() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "CHAR_LENGTH".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string())),
-            ast::Expression::Literal(types::SqlValue::Varchar("extra".to_string())),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row);
-    assert!(result.is_err());
-}
-
-#[test]
-fn test_char_length_wrong_type() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "CHAR_LENGTH".to_string(),
-        args: vec![ast::Expression::Literal(types::SqlValue::Integer(123))],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row);
-    assert!(result.is_err());
-}
-
-#[test]
-fn test_octet_length_wrong_arg_count() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "OCTET_LENGTH".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("hello".to_string())),
-            ast::Expression::Literal(types::SqlValue::Varchar("extra".to_string())),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row);
-    assert!(result.is_err());
-}
-
-#[test]
-fn test_octet_length_wrong_type() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "OCTET_LENGTH".to_string(),
-        args: vec![ast::Expression::Literal(types::SqlValue::Integer(123))],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row);
-    assert!(result.is_err());
-}
-
-#[test]
-fn test_lower_wrong_arg_count() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "LOWER".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Varchar("TEST".to_string())),
-            ast::Expression::Literal(types::SqlValue::Varchar("extra".to_string())),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row);
-    assert!(result.is_err());
-}
-
-#[test]
-fn test_lower_wrong_type() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "LOWER".to_string(),
-        args: vec![ast::Expression::Literal(types::SqlValue::Integer(123))],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row);
-    assert!(result.is_err());
-}
-
-#[test]
-fn test_lower_character_type() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "LOWER".to_string(),
-        args: vec![ast::Expression::Literal(types::SqlValue::Character("TEST".to_string()))],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Varchar("test".to_string()));
-}
-
-// Additional edge case: Character type inputs
-#[test]
-fn test_char_length_character_type() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "CHAR_LENGTH".to_string(),
-        args: vec![ast::Expression::Literal(types::SqlValue::Character("test".to_string()))],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Integer(4));
-}
-
-#[test]
-fn test_octet_length_character_type() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "OCTET_LENGTH".to_string(),
-        args: vec![ast::Expression::Literal(types::SqlValue::Character("café".to_string()))],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Integer(5));
-}
-
-#[test]
-fn test_concat_character_type() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "CONCAT".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Character("hello".to_string())),
-            ast::Expression::Literal(types::SqlValue::Character(" ".to_string())),
-            ast::Expression::Literal(types::SqlValue::Character("world".to_string())),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Varchar("hello world".to_string()));
-}
-
-#[test]
-fn test_position_character_type() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "POSITION".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Character("lo".to_string())),
-            ast::Expression::Literal(types::SqlValue::Character("hello".to_string())),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Integer(4));
-}
-
-#[test]
-fn test_replace_character_type() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "REPLACE".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Character("hello".to_string())),
-            ast::Expression::Literal(types::SqlValue::Character("l".to_string())),
-            ast::Expression::Literal(types::SqlValue::Character("r".to_string())),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Varchar("herro".to_string()));
-}
-
-#[test]
-fn test_reverse_character_type() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "REVERSE".to_string(),
-        args: vec![ast::Expression::Literal(types::SqlValue::Character("test".to_string()))],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Varchar("tset".to_string()));
-}
-
-#[test]
-fn test_left_character_type() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "LEFT".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Character("hello".to_string())),
-            ast::Expression::Literal(types::SqlValue::Integer(3)),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Varchar("hel".to_string()));
-}
-
-#[test]
-fn test_right_character_type() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "RIGHT".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Character("hello".to_string())),
-            ast::Expression::Literal(types::SqlValue::Integer(3)),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Varchar("llo".to_string()));
-}
-
-#[test]
-fn test_instr_character_type() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "INSTR".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Character("hello".to_string())),
-            ast::Expression::Literal(types::SqlValue::Character("ll".to_string())),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Integer(3));
-}
-
-#[test]
-fn test_locate_character_type() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "LOCATE".to_string(),
-        args: vec![
-            ast::Expression::Literal(types::SqlValue::Character("ll".to_string())),
-            ast::Expression::Literal(types::SqlValue::Character("hello".to_string())),
-        ],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Integer(3));
-}
-
-#[test]
-fn test_length_character_type() {
-    let (evaluator, row) = create_test_evaluator();
-    let expr = ast::Expression::Function {
-        name: "LENGTH".to_string(),
-        args: vec![ast::Expression::Literal(types::SqlValue::Character("test".to_string()))],
-        character_unit: None,
-    };
-    let result = evaluator.eval(&expr, &row).unwrap();
-    assert_eq!(result, types::SqlValue::Integer(4));
-}
